Move `LocationInfoCache` to the `service` package
