+++ android/src/main/kotlin/net/mullvad/mullvadvpn/service/ServiceInstance.kt
@@ -6,5 +6,6 @@ import net.mullvad.talpid.ConnectivityListener
 data class ServiceInstance(
     val daemon: MullvadDaemon,
     val connectionProxy: ConnectionProxy,
-    val connectivityListener: ConnectivityListener
+    val connectivityListener: ConnectivityListener,
+    val locationInfoCache: LocationInfoCache
 )
